chore: move validations into own file
